backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Fast API server to handle frontend and AI requests"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "sqlalchemy>=2.0.0",
    "bcrypt>=4.0.0",
    "pydantic>=2.0.0",
    "motor>=3.7.1",
    "pycryptodome>=3.20.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "slowapi>=0.1.9",
    "sse-starlette>=1.8.0",
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
    "mcp>=1.0.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4",
    "pdfplumber>=0.10.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "leann; sys_platform != 'win32'",
    "faiss-cpu>=1.7.0; sys_platform == 'win32'",
    "sentence-transformers>=2.2.0; sys_platform == 'win32'",
]
